api/requirements.txt
azure-functions
pyodbc
pydantic
azure-identity
python-dotenv